Move CachedSignals next to Detect in signals.go
The caching logic (cachedOnce/cachedResult, CachedSignals,
resetCachedForTest) belongs with the rest of the Signals/Detect definitions
rather than in transport.go, which is only about the RoundTripper wrapper.
