+++ pkg/clientsignals/signals_test.go
@@ -27,3 +27,23 @@ func TestDetect_AgentAndSourceAreBothEmptyOrBothSet(t *testing.T) {
 		t.Fatalf("Agent and AgentSource must be empty together or set together, got agent=%q source=%q", s.Agent, s.AgentSource)
 	}
 }
+
+func TestCachedSignals_ComputesOnce(t *testing.T) {
+	resetCachedForTest()
+	t.Cleanup(resetCachedForTest)
+
+	t.Setenv("FLY_INVOKED_BY", "cached-tool")
+	first := CachedSignals()
+
+	// Changing the environment after the first call must not affect the
+	// already-cached result.
+	t.Setenv("FLY_INVOKED_BY", "different-tool")
+	second := CachedSignals()
+
+	if first != second {
+		t.Fatalf("expected CachedSignals to return the same cached value on repeated calls, got %+v then %+v", first, second)
+	}
+	if second.Agent != "cached-tool" {
+		t.Fatalf("expected cached value to reflect the environment at the first call, got agent=%q", second.Agent)
+	}
+}
