+++ pkg/clientsignals/transport.go
@@ -4,33 +4,7 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
-	"sync"
 )
-
-var (
-	cachedOnce   sync.Once
-	cachedResult Signals
-)
-
-// CachedSignals returns the process-wide signals, computed once via Detect
-// and cached for the lifetime of the process. Detection involves a
-// parent-process lookup and environment scanning, so callers should fetch
-// this once (e.g. at client-construction time) and reuse the result rather
-// than calling it per request.
-func CachedSignals() Signals {
-	cachedOnce.Do(func() {
-		cachedResult = Detect()
-	})
-
-	return cachedResult
-}
-
-// resetCachedForTest clears the cached signals so tests can exercise Detect
-// against a freshly modified environment. Only for use in this package's
-// own tests.
-func resetCachedForTest() {
-	cachedOnce = sync.Once{}
-}
 
 // userAgentSuffix returns the human-readable
 // "(interactive=...; parent=...; agent=...)" token to append to a
